refactor: 금액 투입 섹션 컴포넌트 분리
